player_api/requirements.txt
flet
mutagen
python-vlc
